requirements.txt
# ============================================================
# AI Road Quality Monitoring System - Requirements
# Install with: pip install -r requirements.txt
# ============================================================

# Core Deep Learning - YOLOv8
ultralytics>=8.0.0          # YOLOv8 framework (includes PyTorch)

# Computer Vision
opencv-python>=4.8.0         # Real-time video capture & processing
Pillow>=10.0.0               # Image manipulation

# Data Science & Evaluation
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0          # Baseline ML methods
matplotlib>=3.7.0            # Plotting metrics
seaborn>=0.12.0              # Advanced visualization

# Dashboard
streamlit>=1.28.0            # Web dashboard framework
folium>=0.14.0               # Interactive maps
streamlit-folium>=0.15.0     # Folium in Streamlit
streamlit-webrtc>=0.47.0     # Cloud-ready webcam streaming

# GPS Simulation & Real GPS
gpsd-py3>=0.3.0              # Real GPS daemon client (optional)
geopy>=2.4.0                 # GPS / geocoding utilities

# Report Generation
reportlab>=4.0.0             # PDF generation
fpdf2>=2.7.0                 # Alternative PDF library

# Dataset Download & Management
roboflow>=1.1.0              # Roboflow dataset API
requests>=2.31.0
tqdm>=4.66.0                 # Progress bars

# Utilities
python-dotenv>=1.0.0         # Environment variables
loguru>=0.7.0                # Better logging
PyYAML>=6.0                  # YAML config files
